move engine utilities to deprecated namespace
