+++ conftest.py
@@ -11,7 +11,7 @@ from sqlalchemy import MetaData, text, Table, select, or_
 from sqlalchemy.exc import OperationalError
 from sqlalchemy_utils import database_exists, create_database, drop_database
 
-from db.engine import add_custom_types_to_ischema_names, create_engine as sa_create_engine
+from db.deprecated.engine import add_custom_types_to_ischema_names, create_engine as sa_create_engine
 from db.sql import install as sql_install
 from db.utils import get_pg_catalog_table
 from db.metadata import get_empty_metadata
